Solution to Find the Length of a String

+++ Find the Length of a String@@676c039d80e1714600f67643/code.c
@@ -4,6 +4,6 @@
 int main() {
     char ch[100];
     scanf("%s",ch);
-    printf("%d",strlen(ch));
+    printf("%d",(int)strlen(ch));
     return 0;
 }
